Do not pre-process test fixtures with eslint --fix since it makes our code less reliable.
Pre-processing fixtures with eslint makes our transform way less reliable
since we basically doing 2 transforms each time. In user-land it won't probably the case
and we need to test only our code.

Even if we want to provide re-formatting of the code with eslint for users, this should
probably be a sepaeate task.